+++ packages/firebase-cypress/src/utils/run-cypress.ts
@@ -4,7 +4,7 @@ import { existsSync, readdirSync, unlinkSync } from 'fs';
 import { installedCypressVersion } from './cypress-version';
 import { CypressRunnerSchema } from './cypress-runner.schema';
 import request from './request';
-import * as killPort from 'kill-port';
+import killPort from 'kill-port';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const Cypress = require('cypress');